+++ packages/library/src/index.ts
@@ -1,6 +1,6 @@
-export * from "./types/UInt";
-export * from "./types/UInt32";
-export * from "./types/UInt64";
-export * from "./types/UInt112";
-export * from "./types/UInt224";
-export * from "./types/PrecisionHelper";
+export * from "./math/UInt";
+export * from "./math/UInt32";
+export * from "./math/UInt64";
+export * from "./math/UInt112";
+export * from "./math/UInt224";
+export * from "./math/PrecisionHelper";
