+++ packages/library/src/math/PrecisionHelper.ts
