+++ packages/library/src/math/UInt112.ts
